Refactored the method ConnectionDefinitionParser::_buildOutput()
It was required to break the function into more parts for simplicity reasons. For this purpose a separate method had to be created to compute the template device parameters for each device. These parameters consist of device U position, device type and other parameters provided as placeholders (that are device type dependent): port number, etc.