Updated the timestamp mechanism for output file backup
It was required that the current date/time is used as timestamp instead of the last modified date/time of the file. On the one hand it looks more logical as it is connected to the actual move operation. On the other hand this helps avoid a compilation error on Linux