Created virtual getter for retrieving the required number of device port input parameters
It was required to remove the initialization of the DevicePort variable representing the required number of input parameters from class constructor in order to improve the extendability of the derived classes. Instead it has been chosen that each derived class communicates the required value to base DevicePort class via getter. This approach has more sense as this member variable is an attribute whose value is determined intrinsically and not from an outer (non-device-port) class