Moved device ports objects ownership to base Parser class
It was required that the Parser class manages these objects as they are registered and activated here. For this reason a parseInputData() method was implemented in the ISubParser interface class

Also some shared pointers had to be converted to raw pointers due to some issues encountered. They might be reconverted to smart pointers in a future commit