Updated the parsing mechanism of switch ports
It was required to trigger an invalid port type (and not port number) error in case the user entered the '-' character as port type input. This was needed because when this character is entered as port type then the management port number (input 'm' for port number) is the only valid input. This combination is however valid only when the device has a management port. For those who don't entering '-' should result in invalid port type error which is written to the labelling table