Renamed the example files
It was required to add two identifiers as prefixes: example number (identifying all connections from an example rack) and connection types (data/management/power). This was needed for easier identification of what the example represents which might be useful for the user when running some tests on the application. Also it is not possible on Windows to open two files with the same name and this renaming might be helpful in this respect. To be noted is that the application naming should be used when running the code: connectiondefinitions.csv, connectioninput.csv, labellingtable.csv. This means the user should first rename the example file before running the app